new homepage! and some animation stuff, and themes fixing
