+++ package.json
@@ -22,6 +22,7 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "postcss-import": "^16.1.0",
     "tailwindcss": "^4",
     "typescript": "^5"
   }
