+++ postcss.config.mjs
@@ -2,4 +2,12 @@ const config = {
   plugins: ["@tailwindcss/postcss"],
 };
 
+
 export default config;
+// module.exports = {
+//   plugins: {
+//     'postcss-import': {},
+//     tailwindcss: {},
+//     autoprefixer: {},
+//   }
+// }
