+++ src/app/layout.tsx
@@ -2,7 +2,8 @@ import "./globals.css";
 import { Themes } from "@/lib/Themes";
 import Link from "next/link";
 
-import { Noto_Sans_Display, Noto_Sans_Mono } from "next/font/google";
+import { Noto_Sans_Display, Noto_Sans_Mono, Raleway } from "next/font/google";
+import { GIcon } from "@/lib/utils";
 
 const NotoSansDisplay = Noto_Sans_Display({
   variable: "--font-nsd",
@@ -13,6 +14,11 @@ const NotoSansMono = Noto_Sans_Mono({
   variable: "--font-nsm",
   subsets: ["latin"],
 });
+
+const RFont = Raleway({
+  variable: "--font-raleway",
+  subsets: ["latin"]
+})
 
 export default function RootLayout({children}: Readonly<{children:React.ReactNode}>) {
   return (
@@ -26,13 +32,14 @@ export default function RootLayout({children}: Readonly<{children:React.ReactNod
         <meta property="og:description" content="Blueprint editor and valuator for ProDSA Services, a subsidiary of BetaOS Services."/>
         <meta property="og:image" content="https://prodsatools.vercel.app/icon.png"/>
       </head>
-      <body className={`${NotoSansDisplay.variable} ${NotoSansMono.variable} antialiased`}>
+      <body className={`${NotoSansDisplay.variable} ${NotoSansMono.variable} ${RFont.variable} antialiased`}>
         {children}
         <footer className={`${Themes.BLUE.textCls} p-3 flex gap-2 flex-wrap justify-center`}>
-          <Link href="/valuate">Ship cost calculator</Link>
-          <Link href="/rates">ProDSA Rates</Link>
-          <Link href="/">ProDSA PrecisionEdit Tools</Link>
-          <Link href="//dsc.gg/ProDSA" target="_blank" className={Themes.GREEN.textCls}>Order ships from ProDSA Services today!</Link>
+          <Link className="blue active " href="/"><GIcon theme={Themes.BLUE}>home</GIcon></Link>
+          <Link className="blue active" href="/valuate">Ship cost calculator</Link>
+          <Link className="blue active" href="/rates">ProDSA Rates</Link>
+          <Link className="blue active " href="/editor">ProDSA PrecisionEdit Tools</Link>
+          <Link className="green text active" href="//dsc.gg/ProDSA" target="_blank">Order ships from ProDSA Services today!</Link>
           <span>Site design by ProDSA Services</span>
         </footer>
       </body>
