+++ src/app/themes.css
@@ -1,0 +1,64 @@
+@import "tailwindcss";
+
+/* export const Themes: { [x: string]: ColourTheme; } = {
+  RED: {
+    textCls: "text-red-400", activeCls: "active:text-red-300",
+    hoverCls: "hover:text-red-500 hover:!bg-red-200 focus:text-red-500 focus:bg-red-200",
+    bgMain: "bg-red-100", bgLight: "bg-red-50", bgStrong: "bg-red-200"
+  },
+  GREEN: {
+    textCls: "text-lime-500", activeCls: "active:text-lime-300",
+    hoverCls: "hover:text-lime-500 hover:!bg-lime-200 focus:text-lime-500 focus:bg-lime-100",
+    bgMain: "bg-lime-100", bgLight: "bg-lime-50", bgStrong: "bg-lime-200"
+  },
+  BLUE: {
+    textCls: "text-blue-400", activeCls: "active:text-blue-300",
+    hoverCls: "hover:text-blue-500 hover:!bg-blue-200 focus:text-blue-500 focus:bg-blue-100",
+    bgMain: "bg-blue-100", bgLight: "bg-blue-50", bgStrong: "bg-blue-200"
+  },
+  GREY: {
+    textCls: "text-gray-400", activeCls: "active:text-gray-300",
+    hoverCls: "hover:text-gray-500 hover:!bg-gray-200 focus:text-gray-500 focus:bg-gray-100",
+    bgMain: "bg-gray-100", bgLight: "bg-gray-50", bgStrong: "bg-gray-200"
+  }
+}; */
+
+
+/* @theme {
+  --text: text-red-400;
+  --active: active:text-red-300;
+  --hover: hover:text-red-500 hover:bg-red-200 focus:text-red-500 focus:bg-red-200;
+  --bgMain: bg-red-100;
+  --bgLight: bg-red-50;
+  --bgStrong: bg-red-200;
+} */
+
+
+.red.text { @apply text-red-400; }
+.red.active { @apply active:text-red-300; }
+.red.hover { @apply hover:text-red-500 hover:bg-red-200 focus:text-red-500 focus:bg-red-200 transition-colors; }
+.red.bgMain { @apply bg-red-100; }
+.red.bgLight { @apply bg-red-50; }
+.red.bgStrong {@apply bg-red-200; }
+
+.blue.text { @apply text-blue-400; }
+.blue.active { @apply active:text-blue-300; }
+.blue.hover { @apply hover:text-blue-500 hover:bg-blue-200 focus:text-blue-500 focus:bg-blue-200 transition-colors; }
+.blue.bgMain { @apply bg-blue-100; }
+.blue.bgLight { @apply bg-blue-50; }
+.blue.bgStrong {@apply bg-blue-200; }
+
+.green.text { @apply text-lime-400; }
+.green.active { @apply active:text-lime-300; }
+.green.hover { @apply hover:text-lime-500 hover:bg-lime-200 focus:text-lime-500 focus:bg-lime-200 transition-colors; }
+.green.bgMain { @apply bg-lime-100; }
+.green.bgLight { @apply bg-lime-50; }
+.green.bgStrong {@apply bg-lime-200; }
+
+.grey.text { @apply text-gray-400; }
+.grey.active { @apply active:text-gray-300; }
+.grey.hover { @apply hover:text-gray-500 hover:bg-gray-200 focus:text-gray-500 focus:bg-gray-200 transition-colors; }
+.grey.bgMain { @apply bg-gray-100; }
+.grey.bgLight { @apply bg-gray-50; }
+.grey.bgStrong {@apply bg-gray-200; }
+
